Changes according review comments (#1008).
