+++ scripts/tempesta.sh
@@ -34,7 +34,6 @@ script_path="$(dirname $0)"
 tdb_path=${TDB_PATH:="$TFW_ROOT/tempesta_db/core"}
 tfw_path=${TFW_PATH:="$TFW_ROOT/tempesta_fw"}
 tls_path=${TLS_PATH:="$TFW_ROOT/tls"}
-class_path=${TFW_CLS_PATH:="$tfw_path/classifier/"}
 tfw_cfg_path=${TFW_CFG_PATH:="$TFW_ROOT/etc/tempesta_fw.conf"}
 sched_path=${TFW_SCHED_PATH:="$tfw_path/sched/"}
 sched_ko_files=($(ls $sched_path/tfw_sched_*.ko))
@@ -43,7 +42,6 @@ tls_mod=tempesta_tls
 tdb_mod=tempesta_db
 tfw_mod=tempesta_fw
 tfw_sched_mod=tfw_sched_$sched
-frang_mod="tfw_frang"
 declare -r LONG_OPTS="help,load,unload,start,stop,restart,reload"
 
 declare devs=$(ip addr show up | awk '/^[0-9]+/ { sub(/:/, "", $2); print $2}')
@@ -112,12 +110,6 @@ load_modules()
 		load_one_module "$ko_file" ||
 			error "cannot load tempesta scheduler module"
 	done
-
-	if grep -q -E "^\s*frang_limits" $tfw_cfg_path; then
-		echo "Load Frang"
-		load_one_module "$class_path/$frang_mod.ko" ||
-			error "cannot load $frang_mod module"
-	fi
 }
 
 unload_modules()
@@ -128,7 +120,6 @@ unload_modules()
 		rmmod $(basename "${ko_file%.ko}")
 	done
 
-	[ "`lsmod | grep \"\<$frang_mod\>\"`" ] && rmmod $frang_mod
 	rmmod $tfw_mod
 	rmmod $tdb_mod
 	rmmod $tls_mod
