+++ tempesta_fw/http.h
@@ -584,6 +584,12 @@ tfw_current_timestamp(void)
 	return ts.tv_sec;
 }
 
+static inline int
+tfw_http_resp_code_range(const int n)
+{
+	return n <= HTTP_CODE_MAX && n >= HTTP_CODE_MIN;
+}
+
 typedef void (*tfw_http_cache_cb_t)(TfwHttpMsg *);
 
 /* Internal (parser) HTTP functions. */
@@ -613,6 +619,5 @@ void tfw_http_send_resp(TfwHttpReq *req, int status, const char *reason);
 
 /* Helper functions */
 char *tfw_http_msg_body_dup(const char *filename, size_t *len);
-int tfw_http_resp_code_range(const int n);
 
 #endif /* __TFW_HTTP_H__ */
