+++ tempesta_fw/main.c
@@ -126,7 +126,6 @@ tfw_cleanup(struct list_head *mod_list)
 
 	if (!tfw_runstate_is_reconfig())
 		tfw_sg_wait_release();
-
 	TFW_LOG("New configuration is cleaned.\n");
 }
 
@@ -398,13 +397,15 @@ tfw_init(void)
 	DO_INIT(apm);
 	DO_INIT(vhost);
 
-	DO_INIT(classifier);
-	DO_INIT(filter);
-	DO_INIT(cache);
-
-	/* Register TLS before HTTP, so HTTP FSM can register TLS hooks. */
+	/*
+	 * Register in order TLS -> HTTP -> limits, for correct
+	 * registration of FSM hooks.
+	 */
 	DO_INIT(tls);
 	DO_INIT(http);
+	DO_INIT(http_limits);
+	DO_INIT(filter);
+	DO_INIT(cache);
 	DO_INIT(http_sess);
 
 	DO_INIT(sync_socket);
